Updated the consensus mechanism such that relevant public keys, signatures and hashes are used.

+++ src/main/java/nl/hanze/ec/node/API.java
@@ -128,7 +128,7 @@ public class API implements Runnable {
         get("/balances", (request, response) -> {
             response.type("application/json");
 
-            int amount = 0;
+            float amount = 0;
             if (request.queryParams().size() == 0) {
                 return new Gson().toJson(
                         new StandardResponse(StatusResponse.SUCCESS, new Gson()
